# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Huerise Daylight Alarm API",
    "description": "Run one YAML-configured Philips Hue daylight alarm.",
    "version": "2.0.0"
  },
  "paths": {
    "/doctor": {
      "get": {
        "tags": [
          "doctor"
        ],
        "summary": "Doctor",
        "operationId": "doctor",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The configured Hue scene does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The YAML configuration is missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConfigurationErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    }
                  ],
                  "title": "Response 422 Doctor"
                }
              }
            }
          },
          "503": {
            "description": "The Hue Bridge is not configured, reachable, or authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/rooms": {
      "get": {
        "tags": [
          "lighting"
        ],
        "summary": "Rooms",
        "operationId": "listRooms",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoomResponse"
                  },
                  "type": "array",
                  "title": "Response Listrooms"
                }
              }
            }
          },
          "422": {
            "description": "The YAML configuration is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConfigurationErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    }
                  ],
                  "title": "Response 422 Listrooms"
                }
              }
            }
          },
          "503": {
            "description": "The Hue Bridge is not configured, reachable, or authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/scenes": {
      "get": {
        "tags": [
          "lighting"
        ],
        "summary": "Scenes",
        "operationId": "listScenes",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AvailableSceneResponse"
                  },
                  "type": "array",
                  "title": "Response Listscenes"
                }
              }
            }
          },
          "422": {
            "description": "The YAML configuration is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConfigurationErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    }
                  ],
                  "title": "Response 422 Listscenes"
                }
              }
            }
          },
          "503": {
            "description": "The Hue Bridge is not configured, reachable, or authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/hue/bridges": {
      "get": {
        "tags": [
          "hue-setup"
        ],
        "summary": "Discover Bridges",
        "operationId": "listHueBridges",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BridgeResponse"
                  },
                  "type": "array",
                  "title": "Response Listhuebridges"
                }
              }
            }
          },
          "422": {
            "description": "The stored Hue configuration is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConfigurationErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    }
                  ],
                  "title": "Response 422 Listhuebridges"
                }
              }
            }
          },
          "503": {
            "description": "Hue Bridge discovery is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/hue/bridge": {
      "get": {
        "tags": [
          "hue-setup"
        ],
        "summary": "Bridge Status",
        "operationId": "getHueBridge",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "The stored Hue configuration is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConfigurationErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    }
                  ],
                  "title": "Response 422 Gethuebridge"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      },
      "put": {
        "tags": [
          "hue-setup"
        ],
        "summary": "Select Bridge",
        "operationId": "selectHueBridge",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BridgeSelectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "The selected Hue Bridge was not discovered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Environment overrides make Hue onboarding read-only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The stored Hue configuration is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConfigurationErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    }
                  ],
                  "title": "Response 422 Selecthuebridge"
                }
              }
            }
          },
          "503": {
            "description": "Hue Bridge discovery is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/hue/bridge/register": {
      "post": {
        "tags": [
          "hue-setup"
        ],
        "summary": "Register Bridge",
        "operationId": "registerHueBridge",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingStatusResponse"
                }
              }
            }
          },
          "409": {
            "description": "A Hue Bridge must be selected before registration.\n\nThe Hue Bridge link button was not pressed in time.\n\nEnvironment overrides make Hue onboarding read-only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The stored Hue configuration is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConfigurationErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    }
                  ],
                  "title": "Response 422 Registerhuebridge"
                }
              }
            }
          },
          "503": {
            "description": "Hue Bridge registration is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/daylight-alarm/start": {
      "post": {
        "tags": [
          "daylight-alarm"
        ],
        "summary": "Start",
        "operationId": "startDaylightAlarm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/StartRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmStatusResponse"
                }
              }
            }
          },
          "409": {
            "description": "A daylight alarm is already running.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The configured Hue scene does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The YAML configuration is missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConfigurationErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    }
                  ],
                  "title": "Response 422 Startdaylightalarm"
                }
              }
            }
          },
          "503": {
            "description": "The Hue Bridge is not configured, reachable, or authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/daylight-alarm/stop": {
      "post": {
        "tags": [
          "daylight-alarm"
        ],
        "summary": "Stop",
        "operationId": "stopDaylightAlarm",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AlarmStatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "started",
            "title": "Status",
            "default": "started"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "duration_seconds"
        ],
        "title": "AlarmStatusResponse"
      },
      "AvailableSceneResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "room_id": {
            "type": "string",
            "format": "uuid",
            "title": "Room Id"
          },
          "room_name": {
            "type": "string",
            "title": "Room Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "room_id",
          "room_name"
        ],
        "title": "AvailableSceneResponse"
      },
      "BridgeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "ip_address": {
            "type": "string",
            "title": "Ip Address"
          },
          "selected": {
            "type": "boolean",
            "title": "Selected"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ip_address",
          "selected"
        ],
        "title": "BridgeResponse"
      },
      "BridgeSelectionRequest": {
        "properties": {
          "bridge_id": {
            "type": "string",
            "minLength": 1,
            "title": "Bridge Id"
          }
        },
        "type": "object",
        "required": [
          "bridge_id"
        ],
        "title": "BridgeSelectionRequest"
      },
      "ConfigurationErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ConfigurationIssue"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "issues"
        ],
        "title": "ConfigurationErrorResponse"
      },
      "ConfigurationIssue": {
        "properties": {
          "location": {
            "type": "string",
            "title": "Location"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "location",
          "message",
          "type"
        ],
        "title": "ConfigurationIssue"
      },
      "DoctorCheckResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "name",
          "status"
        ],
        "title": "DoctorCheckResponse"
      },
      "DoctorResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/DoctorCheckResponse"
            },
            "type": "array",
            "title": "Checks"
          }
        },
        "type": "object",
        "required": [
          "status",
          "checks"
        ],
        "title": "DoctorResponse"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OnboardingState": {
        "type": "string",
        "enum": [
          "not_selected",
          "link_button_required",
          "ready"
        ],
        "title": "OnboardingState"
      },
      "OnboardingStatusResponse": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/OnboardingState"
          },
          "bridge_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bridge Id"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "read_only": {
            "type": "boolean",
            "title": "Read Only"
          }
        },
        "type": "object",
        "required": [
          "state",
          "bridge_id",
          "ip_address",
          "read_only"
        ],
        "title": "OnboardingStatusResponse"
      },
      "RoomResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "scenes": {
            "items": {
              "$ref": "#/components/schemas/SceneResponse"
            },
            "type": "array",
            "title": "Scenes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "scenes"
        ],
        "title": "RoomResponse"
      },
      "SceneResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "SceneResponse"
      },
      "StartRequest": {
        "properties": {
          "duration_seconds": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "duration_seconds"
        ],
        "title": "StartRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}